requirements.txt
flask
pytest
coverage
click
psycopg2-binary